requirements.txt
aiohttp>=3.9.0
aiohttp-jinja2>=1.5.0
jinja2>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
hyperliquid-python-sdk>=0.3.0
eth-account>=0.10.0
